工时管理软件git项目，修改派工单实体类的数据类型，完成服务器接口的修改，

+++ app/src/main/java/com/shg/manhourapp/domain/DispatchListItemsViewModel.java
@@ -12,9 +12,9 @@ import java.io.Serializable;
 
 public class DispatchListItemsViewModel implements Serializable{
 
-    public int ManHourActual;
+    public double ManHourActual;
     public String MaterialName;
-    public int Volume;
+    public double Volume;
     public String ShiftName;
     public String ConstructionSiteName;
     public String EquipmentName;
